requirements.txt
python-telegram-bot==20.3
yt-dlp==2024.03.03
ffmpeg-python==0.2.0